Reorganize nullable ToTry under Try/ slice; switch tests to FsCheck
Feature-slice layout: the nullable-to-Try extensions belong to the Try
feature (the output type defines the slice, not the input), so the file
moves from Extensions/NullableExtensions.cs to Try/TryExtensions.cs. The
class is renamed to TryExtensions to match. To free that name, the legacy
TryExtensions in Try_Old is renamed to TryExtensions_Old (mirrors the
pattern used for StringExtensions).

Tests move to Try/TryExtensionsTests.cs. The namespace stays flat at
StrongTypes.Tests — nesting into StrongTypes.Tests.Try would shadow the
StrongTypes.Try class from sibling _Old tests.

Switch the two ToTry tests to FsCheck property tests:
- One Arbitrary<NonEmptyString?> generator producing both null and
  populated values, defined in a nested Generators class and registered
  via [Properties(Arbitrary = ...)].
- ParseError enum is two-valued with Missing = 1 so 'not default' is a
  real assertion about propagation rather than an accident of zero-init.
- ToTry_ReferenceType covers the class overload; ToTry_ValueType covers
  the struct overload with int?.

CLAUDE.md: document the feature-slice folder convention and the
flat-namespace rule, with the ToTry case as the worked example.
