+++ src/StrongTypes/Try/TryExtensions.cs
@@ -5,10 +5,9 @@ using System;
 namespace StrongTypes;
 
 /// <summary>
-/// Bridges between nullable reference types and the result-shaped types in
-/// this library.
+/// Extensions that produce <see cref="Try{A, E}"/> values from other shapes.
 /// </summary>
-public static class NullableExtensions
+public static class TryExtensions
 {
     /// <summary>
     /// Turns a nullable reference into a <see cref="Try{A, E}"/>: a non-null
